Check before unlink for a null string

+++ src/libsrcml/srcml_transform.cpp
@@ -187,7 +187,7 @@ int srcml_apply_transforms(srcml_archive* iarchive, srcml_archive* oarchive) {
  
       close(transform_fd);
       if(i != 0) xmlFreeParserInputBuffer(pinput);
-      unlink(last_transform_filename);
+      if(last_transform_filename)  unlink(last_transform_filename);
       free((void *)last_transform_filename);
   
       return SRCML_STATUS_ERROR;
@@ -198,7 +198,7 @@ int srcml_apply_transforms(srcml_archive* iarchive, srcml_archive* oarchive) {
     free((void *)last_transform_filename);
     last_transform_filename = transform_filename;
     if(error == SRCML_STATUS_ERROR) {
-      unlink(last_transform_filename);
+      if(last_transform_filename) unlink(last_transform_filename);
       free((void *)last_transform_filename);
       return error;
     }
@@ -219,7 +219,7 @@ int srcml_apply_transforms(srcml_archive* iarchive, srcml_archive* oarchive) {
 
   srcml_close_archive(tmp_archive);
   srcml_free_archive(tmp_archive);
-  unlink(last_transform_filename);
+  if(last_transform_filename) unlink(last_transform_filename);
   free((void *)last_transform_filename);
 
   iarchive->transformations.clear();
